Fix logging, despaghettify a bit, change ports to ints

+++ UnchainedLauncherGUI/src/JsonModels/ServerSettings.cs
@@ -6,10 +6,10 @@ namespace UnchainedLauncher.GUI.JsonModels {
         [property: JsonProperty("server_description")] string? ServerDescription,
         [property: JsonProperty("server_password")] string? ServerPassword,
         [property: JsonProperty("selected_map")] string? SelectedMap,
-        [property: JsonProperty("game_port")] short? GamePort,
-        [property: JsonProperty("rcon_port")] short? RconPort,
-        [property: JsonProperty("a2s_port")] short? A2sPort,
-        [property: JsonProperty("ping_port")] short? PingPort,
+        [property: JsonProperty("game_port")] int? GamePort,
+        [property: JsonProperty("rcon_port")] int? RconPort,
+        [property: JsonProperty("a2s_port")] int? A2sPort,
+        [property: JsonProperty("ping_port")] int? PingPort,
         [property: JsonProperty("show_in_server_browser")] bool? ShowInServerBrowser
     );
 }
